change unsigned long to long long in philo one

+++ philo_one/incs/philosophers.h
@@ -6,7 +6,7 @@
 /*   By: mroux <mroux@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/10 23:33:24 by mroux             #+#    #+#             */
-/*   Updated: 2021/03/24 22:53:42 by mroux            ###   ########.fr       */
+/*   Updated: 2021/03/30 19:48:14 by mroux            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ typedef struct		s_philo
 	int				philo_number;
 	struct timeval	last_lunch;
 	int				meals;
-	unsigned long	started_at;
+	long long	started_at;
 	pthread_mutex_t	*forks[2];
 	int				status;
 	void			*ret;
@@ -35,9 +35,9 @@ typedef struct		s_philo
 typedef struct		s_global
 {
 	t_philo			*philos;
-	unsigned long 	time_to_die;
-	unsigned long	time_to_eat;
-	unsigned long	time_to_sleep;
+	long long 	time_to_die;
+	long long	time_to_eat;
+	long long	time_to_sleep;
 	int				number_of_philos;
 	int				number_of_meals;
 	pthread_mutex_t	mutex_print;
@@ -70,10 +70,10 @@ void				think(t_philo *philo, t_global *gl);
 *	Utils
 */
 
-void				ft_usleep(unsigned long n);
+void				ft_usleep(long long n);
 int					ft_atoul(char const *str);
-void				print_death(int philo_number, unsigned long death_time, t_global *gl);
-void				print_end(int max_meal, unsigned long end_time, t_global *gl);
+void				print_death(int philo_number, long long death_time, t_global *gl);
+void				print_end(int max_meal, long long end_time, t_global *gl);
 
 /*
 *	Prints
@@ -85,9 +85,9 @@ void				ft_putnbr_fd(int n, int fd);
 *	Time
 */
 
-unsigned long		timeval_to_ms(struct timeval *tp);
-unsigned long		get_diff_in_ms(struct timeval *tp1, struct timeval *tp2);
-unsigned long		get_relative_time_in_ms(unsigned long ref);
+long long		timeval_to_ms(struct timeval *tp);
+long long		get_diff_in_ms(struct timeval *tp1, struct timeval *tp2);
+long long		get_relative_time_in_ms(long long ref);
 
 /*
 *	Parsing
@@ -95,9 +95,9 @@ unsigned long		get_relative_time_in_ms(unsigned long ref);
 
 int					check_vars(int argc, char *argv[]);
 int					extract_number_of_philosophers(int argc, char *argv[]);
-unsigned long		extract_time_to_die(int argc, char *argv[]);
-unsigned long		extract_time_to_eat(int argc, char *argv[]);
-unsigned long		extract_time_to_sleep(int argc, char *argv[]);
-unsigned long		extract_number_of_meals(int argc, char *argv[]);
+long long		extract_time_to_die(int argc, char *argv[]);
+long long		extract_time_to_eat(int argc, char *argv[]);
+long long		extract_time_to_sleep(int argc, char *argv[]);
+long long		extract_number_of_meals(int argc, char *argv[]);
 
 #endif
