+++ philo_one/srcs/time.c
@@ -6,22 +6,22 @@
 /*   By: mroux <mroux@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/11 00:48:07 by mroux             #+#    #+#             */
-/*   Updated: 2021/03/11 01:59:15 by mroux            ###   ########.fr       */
+/*   Updated: 2021/03/30 19:51:01 by mroux            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philosophers.h"
 
-unsigned long	timeval_to_ms(struct timeval *tp)
+long long	timeval_to_ms(struct timeval *tp)
 {
-	return (tp->tv_sec * 1000 + tp->tv_usec / 1000);
+	return ((long long)((long long)tp->tv_sec * 1000 + (long long)tp->tv_usec / 1000));
 }
-unsigned long	get_diff_in_ms(struct timeval *tp1, struct timeval *tp2)
+long long	get_diff_in_ms(struct timeval *tp1, struct timeval *tp2)
 {
 	return (timeval_to_ms(tp2) - timeval_to_ms(tp1));
 }
 
-unsigned long	get_relative_time_in_ms(unsigned long ref)
+long long	get_relative_time_in_ms(long long ref)
 {
 	struct timeval	tp;
